+++ src/components/molecules/EarthEditor/TagPane/common.ts
@@ -1,0 +1,5 @@
+import { Tag as TagType } from "./TagGroup";
+
+export type Tag = TagType;
+
+export const DEFAULT_TAG_ID = "default";
